Add sig file to avoid re-configure

+++ docker-entrypoint.sh
@@ -1,38 +1,42 @@
 #!/bin/bash
 
-# Set environment variables
-DATASOURCE=java:jboss/datasources/${DB_NAME}DS
+if [[ ! -f $JBOSS_HOME/.setup ]]; then
+    # Set environment variables
+    DATASOURCE=java:jboss/datasources/${DB_NAME}DS
 
-# Setup WildFly admin user
-echo "=> Add WildFly administrator"
-$JBOSS_HOME/bin/add-user.sh --silent=true $WILDFLY_USER $WILDFLY_PASS
+    # Setup WildFly admin user
+    echo "=> Add WildFly administrator"
+    $JBOSS_HOME/bin/add-user.sh --silent=true $WILDFLY_USER $WILDFLY_PASS
 
-# Configure datasource
-echo "=> Create datasource: '${DATASOURCE}'"
-sed -i "/<drivers>/i\\
-                <datasource jndi-name=\"${DATASOURCE}\" pool-name=\"${DB_NAME}Pool\" enabled=\"true\" jta=\"true\" use-java-context=\"true\">\\
-                    <connection-url>jdbc:mysql://${DB_URI}/${DB_NAME}</connection-url>\\
-                    <driver>mysql</driver>\\
-                    <pool>\\
-                        <max-pool-size>25</max-pool-size>\\
-                    </pool>\\
-                    <security>\\
-                        <user-name>${DB_USER}</user-name>\\
-                        <password>${DB_PASS}</password>\\
-                    </security>\\
-                </datasource>\\
-" $JBOSS_HOME/standalone/configuration/standalone.xml
-sed -i "/<\\/drivers>/i\\
-                    <driver name=\"mysql\" module=\"com.mysql\">\\
-                        <xa-datasource-class>com.mysql.jdbc.jdbc2.optional.MysqlXADataSource</xa-datasource-class>\\
-                    </driver>
-" $JBOSS_HOME/standalone/configuration/standalone.xml
+    # Configure datasource
+    echo "=> Create datasource: '${DATASOURCE}'"
+    sed -i "/<drivers>/i\\
+                    <datasource jndi-name=\"${DATASOURCE}\" pool-name=\"${DB_NAME}Pool\" enabled=\"true\" jta=\"true\" use-java-context=\"true\">\\
+                        <connection-url>jdbc:mysql://${DB_URI}/${DB_NAME}</connection-url>\\
+                        <driver>mysql</driver>\\
+                        <pool>\\
+                            <max-pool-size>25</max-pool-size>\\
+                        </pool>\\
+                        <security>\\
+                            <user-name>${DB_USER}</user-name>\\
+                            <password>${DB_PASS}</password>\\
+                        </security>\\
+                    </datasource>\\
+    " $JBOSS_HOME/standalone/configuration/standalone.xml
+    sed -i "/<\\/drivers>/i\\
+                        <driver name=\"mysql\" module=\"com.mysql\">\\
+                            <xa-datasource-class>com.mysql.jdbc.jdbc2.optional.MysqlXADataSource</xa-datasource-class>\\
+                        </driver>
+    " $JBOSS_HOME/standalone/configuration/standalone.xml
 
-echo "=> Clean up"
-## FIX for Error: WFLYCTL0056: Could not rename /opt/jboss/wildfly/standalone/configuration/standalone_xml_history/current...
-rm -rf $JBOSS_HOME/standalone/configuration/standalone_xml_history/* \
-       $JBOSS_HOME/standalone/log/*
-unset WILDFLY_USER WILDFLY_PASS DB_NAME DB_USER DB_PASS DATASOURCE
+    echo "=> Clean up"
+    ## FIX for Error: WFLYCTL0056: Could not rename /opt/jboss/wildfly/standalone/configuration/standalone_xml_history/current...
+    rm -rf $JBOSS_HOME/standalone/configuration/standalone_xml_history/* \
+        $JBOSS_HOME/standalone/log/*
+    unset WILDFLY_USER WILDFLY_PASS DB_NAME DB_USER DB_PASS DATASOURCE
+    
+    touch $JBOSS_HOME/.setup
+fi
 
 if [[ "$WILDFLY_DEBUG" = "true" ]]; then
     echo "=> Enable debug mode"
